requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
pyyaml>=6.0.0
